move GC code into core/repo, and add sharness test
